==> Part4_Predictive_Mixed_Model_Analysis.Rmd <==
---
title: "Part4: Predictive Model"
author: "Ilse van Beelen"
date: "11/12/2019"
output: html_document
---

```{r setup, include=FALSE, echo=TRUE, warning=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = "hide")



```

Goal: create a prediction model for the NDI at 52 and 104 weeks, using the NDI, HADS anx & depres at baseline. 

# Set up

```{r, warning=FALSE}

# Load libraries
rm(list = ls())
set.seed(19950306)
library(ggplot2)
library(lme4)
library(tidyverse)
library(nlme)
library(emmeans)
library(car)
library(lattice)
library(merTools)
library(xtable)
library(caret)
library(groupdata2)

# Set layout for all figures
theme <- theme(panel.background = element_blank(),
          panel.grid.major = element_line(colour = "darkgrey", size=0.5),
          panel.grid.minor = element_line(colour = "grey", 
                                          size=.25, 
                                          linetype = "dashed"),
          panel.border = element_blank(),
          axis.line.x = element_line(colour = "black", 
                                     size=0.5, 
                                     lineend = "butt"),
          axis.line.y = element_line(colour = "black", 
                                     size=0.5),
          axis.text=element_text(size=15),
              axis.title=element_text(size=22),
              plot.title = element_text(size = 22),
              strip.text = element_text(size = 15),
              legend.title = element_blank())



df_prediction <- read.csv("Data/data_final_prediction_2019-11-05.csv", sep = ",")



# Center predictors -> makes interpretation intercept easier
# Create new numbering voor de IDs -> for Cross validation
df_prediction <- df_prediction %>% 
  mutate(id_revalue = as.factor(group_indices_(.,.dots=list("id")))) %>%
  mutate(time_fct = as.factor(time_fct)) %>%
  mutate(ndi0_cnt =  ndi0 - mean(ndi0)) %>%
  mutate(hads0_tot_cnt = hads0_tot - mean(hads0_tot))


str(df_prediction)

```

# Missing measurements

We omit visit from the data when NDI and/or HADS is missing

```{r}
df_prediction %>%
  filter_all(any_vars(is.na(.)))

df_prediction_with_na <- df_prediction 

df_prediction <- df_prediction %>%
  drop_na(ndi1_2)

```

# Prediction Model: find best random effects

Fit random intercept and model with random intercept + slope. Both fitted with REML

```{r}

# random intercept
model_final <- nlme::lme(ndi1_2 ~ ndi0_cnt + hads0_tot_cnt + time_fct, method = "REML",
                         random = ~1|id, data = df_prediction)

# random intercept + slope
model_final2 <- nlme::lme(ndi1_2 ~ ndi0_cnt + hads0_tot_cnt + time_fct, method = "REML",
                         random = ~1+time_fct|id, data = df_prediction)


model_final3 <- lmer(ndi1_2 ~ ndi0_cnt + hads0_tot_cnt + time_fct + (1|id) + (0+time_fct|id), 
                     data = df_prediction,REML = F)


anova(model_final, model_final2)

plot(ranef(model_final2, order=T))

df_ranef <- as.data.frame(ranef(model_final2))
plot(sort(df_ranef[,1], df_ranef[,2]))

df_ranef <- df_ranef[order(df_ranef$`(Intercept)`),]

plot(df_ranef$`(Intercept)`, df_ranef$time_fct104)



```


# Prediction Model

We fit the following prediction model, using REML.

$$ NDI_{i1,i2} = \beta_0 + \beta_1 HADS_{i0} + \beta_2 NDI_{i0} + \beta_3 Time_{i1, i2} + b_{i0} + \varepsilon_{i1, i2} \\
i = 1,2,3, \ldots, 109 $$

```{r}

# random intercept
model_final <- nlme::lme(ndi1_2 ~ ndi0_cnt + hads0_tot_cnt + time_fct, method = "REML",
                         random = ~1|id, data = df_prediction)

summary(model_final)
model_final_coef <- data.frame(coef(summary(model_final)))
xtable(model_final_coef)


```

## Calculate Intra-Class Correlation (ICC)

This is the amount of variance explained

```{r}

model_icc <- nlme::lme(ndi1_2 ~ 1, method = "REML",
                         random = ~1|id, data = df_prediction)

summary(model_icc)


# Calculate ICC (amount of var explained = 75 %
15.56856^2 / (15.56856^2 + 8.937615^2)


```


## Plot predictions for four specific patients


Make prediction for four patients with 95 % confidence intervals.

```{r}

# Make prediction for subject L8141003
validation1 <- df_prediction[df_prediction$id == "L8141003",]
training1 <- df_prediction[df_prediction$id != "L8141003", ]

# Make prediction for subject L8105004
validation1 <- df_prediction[df_prediction$id == "L8105004",]
training1 <- df_prediction[df_prediction$id != "L8105004", ]

training.fit <-  lmer(ndi1_2 ~ ndi0 + hads0_tot_cnt + time_fct + (1|id), REML = T,
                      data = training1)


preds1 <- predictInterval(training.fit, newdata = validation1, n.sims = 999)
validation1 <- cbind(validation1, preds1)

# Make prediction for subject L8061024
validation2 <- df_prediction[df_prediction$id == "L8061024",]
training2 <- df_prediction[df_prediction$id != "L8061024",]

training.fit <-  lmer(ndi1_2 ~ ndi0 + hads0_tot + time_fct + (1|id), REML = T,
                      data = training2)

preds2 <- predictInterval(training.fit, newdata = validation2, n.sims = 999)
validation2 <- cbind(validation2, preds2)

# Make predictions for L8043008
validation3 <- df_prediction[df_prediction$id == "L8043008",]
training3 <- df_prediction[df_prediction$id != "L8043008",]

training.fit <-  lmer(ndi1_2 ~ ndi0 + hads0_tot + time_fct + (1|id), REML = T,
                      data = training3)

preds3 <- predictInterval(training.fit, newdata = validation3, n.sims = 999)
validation3 <- cbind(validation3, preds3)

# Make predictions for L8061013
validation4 <- df_prediction[df_prediction$id == "L8061013",]
training4 <- df_prediction[df_prediction$id != "L8061013",]

training.fit <-  lmer(ndi1_2 ~ ndi0 + hads0_tot + time_fct + (1|id), REML = T,
                      data = training4)

preds4 <- predictInterval(training.fit, newdata = validation4, n.sims = 999)
validation4 <- cbind(validation4, preds4)

# Combine predictions in one dataframe and plot the results
predict_4id <- rbind(validation1, validation2, validation3, validation4)

```

Plot the two predictions

```{r}

predict_4id$time <- as.numeric(as.character(predict_4id$time_fct))

predict_4id <- predict_4id %>%
  #add_row(id = "L8141003", ndi1_2 = 44, fit = 44, time = 0, upr = 44, lwr = 44, .before = 1) %>%
  add_row(id = "L8105004", ndi1_2 = 44, fit = 44, time = 0, upr = 44, lwr = 44, .before = 1) %>%
  add_row(id = "L8061024", ndi1_2 = 42, fit = 42, time = 0, upr = 42, lwr = 42, .before =  4) %>%
  add_row(id = "L8043008", ndi1_2 = 36, fit = 36, time = 0, upr = 36, lwr = 36, .before =  7) %>%
  add_row(id = "L8061013", ndi1_2 = 18, fit = 18, time = 0, upr = 18, lwr = 18, .before =  10)

ppredict <- ggplot(data = predict_4id, aes(x = time, y = ndi1_2)) + 
  geom_line(size = 1.5, col = "black") +
  geom_point(size = 3, col = "black") +
  geom_line(aes(x = time, y = fit, col = id), size = 1.5) +
  geom_point(aes(x = time, y = fit, col = id), size = 3, shape = 17) +
  geom_ribbon(aes(ymin= lwr, ymax= upr, col = id, fill = id), alpha=0.3) + theme +
  ylab("NDI") + xlab("Time (weeks)") +  facet_grid(~id) + theme(legend.position = "none")

plot(ppredict)

png("Figures/Predicted_ids.png",width = 15, height = 7, units='in',res=300)
plot(ppredict)
dev.off()



```



## 5 fold Cross validation (CV)


Below we devide the data in 5 folds. The folds are grouped on the IDs, meaning that the measurements of one patient are grouped in one fold. We calculate the Root Mean SQuared Error (RMSE).



```{r}

# Add new column with folds numbers
K <- 5
df_prediction <- fold(df_prediction, k = K, id_col = "id_revalue") %>%
  print(n=Inf)


# Create Loss function: Root Mean Squard Error
Loss <- function(x, y){
  mse <- sum((x-y)^2)/length(x)
  rmse <- sqrt(mse)
  return(rmse)
}


loss <- numeric(K)
list_pred <- c()

for (k in 1:K){
  training <- df_prediction[df_prediction$.folds !=k, ]
  validation <- df_prediction[df_prediction$.folds ==k, ]
  training.fit <-  lme4::lmer(ndi1_2 ~ ndi0 + hads0_anx + hads0_depr + time_fct + (1|id), 
                    data = training, REML = T)
  
  validation.predict <- predict(training.fit, newdata=validation, type='response',
                                 allow.new.levels = T)
  loss[k] <- Loss(validation$ndi1_2, validation.predict)
  
  # Save predictions in list
  list_pred[[k]] <- validation %>%
    dplyr::select(id_revalue, id, ndi1_2, time_fct, .folds) %>%
    mutate(pred = validation.predict)
}


# Show RMSE + SD
loss
RMSE  <- round(mean(loss), digits = 1)
RMSE
round(sd(loss), digits = 1)


# 95 % CI of RMSE
RMSE_se <- sd(loss) / sqrt(100)
lower <- round(mean(loss) - 1.96 * RMSE_se, digits = 1)
upper  <- round(mean(loss) + 1.96 * RMSE_se, digits = 1)
c(lower, upper)

```



## Plot histogram of predictions

Barplot visualization of one fold of the Cross Validation

```{r}
# Create dataframe
predictions <- round(validation.predict, digits = 0)
predictions_not_rounded <- validation.predict
validations <- validation$ndi1_2
subject <- validation$id
timepoint <- validation$time_fct
id_time <- paste(subject, timepoint, sep = "-")

df_hist_pred1 <- data.frame(id_time = id_time, id = subject, time = timepoint, 
                            value = predictions,  pred_not_rounded = predictions_not_rounded,
                            var = "predictions")
df_hist_pred2 <- data.frame(id_time = id_time, id = subject, time = timepoint, 
                            value = validations, pred_not_rounded = predictions_not_rounded,
                            var = "validations")
df_hist_pred <- rbind(df_hist_pred1, df_hist_pred2)

df_hist_pred <- df_hist_pred[order(df_hist_pred$time),]


# Plot predictions and validation together
p_bar <- ggplot(data=df_hist_pred, aes(x=id_time, y=value)) +
  geom_bar(stat="identity", position=position_dodge(), aes(fill = var, colour = var))+
  #geom_text(aes(label=value), vjust=1.6, color="white",
   #         position = position_dodge(3), size=3.5)+
  ylab("NDI") + xlab("Patient ID & visit") + 
  theme + 
  scale_color_manual(values = c('validations' = '#333333', 'predictions' = 'red')) +
  scale_fill_manual(values = c('validations' = '#333333', 'predictions' = 'red')) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  theme(legend.position="bottom", legend.text = element_text( size=15))  #+ facet_grid(~time) + 

plot(p_bar)

png("Figures/Predicted_ids_barplot.png",width = 15, height = 7, units='in',res=300)
plot(p_bar)
dev.off()


```

## Calculate RMSE for 52 and 104 weeks seperate

We have a look at the RMSE for 52 and 104 weeks seperately, because we expected that the predictions for week 52 are more accurate. The difference in RMSE turned about to be neglectable.


```{r}

df_cv <- do.call("rbind", list_pred)

df_cv52 <- df_cv %>%
  filter(time_fct == 52)


df_cv104 <- df_cv %>%
  filter(time_fct == 104)

# calculate rmse per time point
df_cv52 <- df_cv52 %>%
  group_by(.folds) %>%
  mutate(loss = Loss(ndi1_2, pred))


rmse_52 <- unique(df_cv52$loss)
round(mean(rmse_52), 1)
round(sd(rmse_52), 1)


df_cv104 <- df_cv104 %>%
  group_by(.folds) %>%
  mutate(loss = Loss(ndi1_2, pred))

rmse_104 <- unique(df_cv104$loss)
round(mean(rmse_104), 1)
round(sd(rmse_104), 1)

```
