==> Part5_Data_Cleaning_Improve_Predictions.Rmd <==
---
title: 'Part5: data cleaning to improve predictions'
author: "Ilse van Beelen"
date: "2020/03/23"
output: html_document
---


```{r setup, include=FALSE, echo=TRUE, warning=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = "hide")


```

Goal of this file: merge mutiple datasets together. Next part will try to improve prediction model with the new variables.

# Set up

```{r, warning=FALSE}

# Load libraries
rm(list = ls())
set.seed(19950306)

library(tidyverse)
library(haven)


# read in datasets
df_prediction <- read.csv("Data/data_final_prediction_2019-11-05.csv", sep = ",")
dat_raw <- read_sav("Data/Dataset NECK HADS stuk 16-05-2019.sav")
dat_aa <- read_sav("Data/Dataset_AA.sav")
dat_ba <- read_sav("Data/Dataset_BA.sav")
dat_bc <- read_sav("Data/Dataset_BC.sav")
dat_dh <- read_sav("Data/Dataset_Disc_Height.sav")

```

# Prep prediction data (same as older scripts)

```{r}

# Center predictors -> makes interpretation intercept easier
# Create new numbering voor de IDs -> for Cross validation
df_prediction <- df_prediction %>% 
  mutate(id_revalue = as.factor(group_indices_(.,.dots=list("id")))) %>%
  mutate(time_fct = as.factor(time_fct)) %>%
  mutate(ndi0_cnt =  ndi0 - mean(ndi0)) %>%
  mutate(hads0_tot_cnt = hads0_tot - mean(hads0_tot))

# remove missing data
# We omit visit from the data when NDI and/or HADS is missing
df_prediction <- df_prediction %>%
  drop_na(ndi1_2)


```



# Some extra data cleaning

```{r}
# dat_raw: age.ok, gender, smoking, alcohol, weight, length, bmi, duur_klachten
dat1 <- dat_raw %>%
  dplyr::select(PATIENTID, PMOMENTX, AGE.OK, GENDER, Smoking, Alcohol, WEIGHT, LENGTH, 
                BMI, Duur_Klachten) %>%
  dplyr::filter(PMOMENTX == 0) %>%
  dplyr::select(-PMOMENTX)

# dat_aa:
dat2 <- dat_aa %>%
  dplyr::select(ID, IDAA, VINT01A, VINT01W, VINT01J, NEKPIJN, OSTSPOND) 

# dat_ba:
dat3 <- dat_ba %>%
  dplyr::select(ID, IDAA, NEKHFAM) %>%
  drop_na(NEKHFAM)

# dat_bc:
dat4 <- dat_bc %>%
  dplyr::select(ID, IDAA, PMOMENTX, VVAS01, VVAS02, VDLV15A2) %>%
  dplyr::filter(PMOMENTX == -1) %>%
  dplyr::select(-PMOMENTX)

# dat_dh:
# dh = disc height
dat5 <- dat_dh %>%
  dplyr::select(PATIENTID, DHindexlevelBaseline) %>%
  mutate(PATIENTID = paste0("L", PATIENTID)) %>%
  drop_na(DHindexlevelBaseline)


# merge all datasets
# two datasets below have a patientid column
dat15 <- left_join(dat1, dat5, by = c( "PATIENTID"))

# create patientid for datasets below 
dat23 <- full_join(dat2, dat3, by = c("ID", "IDAA"))
dat234 <- full_join(dat23, dat4, by = c("ID", "IDAA"))

dat234 <- dat234 %>%
  mutate(PATIENTID = paste0("L", ID, IDAA)) %>%
  dplyr::select(-ID, -IDAA)



dat_tot <- full_join(dat234, dat15, by = c( "PATIENTID"))
dat_tot <- dat_tot %>% 
  dplyr::select(PATIENTID, everything()) %>%
  rename(id = PATIENTID)

# laste merge
df_all <- left_join(df_prediction, dat_tot, by = "id") # only wants id's who are in df_prediction
df_all <- df_all %>% dplyr::select(id, id_revalue, everything())

```

# Save 

```{r}

write.csv(df_all, "Data/df_prediction_new_2020-03-23.csv")

```
